Bypassクラスを分割する
