+++ lib/procon_bypass_man/bypass/bypass_value.rb
@@ -1,0 +1,6 @@
+class ProconBypassMan::Bypass::BypassValue < Struct.new(:binary)
+  def to_text
+    return unless binary
+    binary.unpack.first
+  end
+end
